+++ jsonargparse/util.py
@@ -156,7 +156,7 @@ def usage_and_exit_error_handler(parser: 'ArgumentParser', message: str) -> None
         parser.exit(2)
 
 
-def _issubclass(cls, class_or_tuple):
+def is_subclass(cls, class_or_tuple):
     """Extension of issubclass that supports non-class argument."""
     return inspect.isclass(cls) and issubclass(cls, class_or_tuple)